backend/pdf/requirements.txt
fastapi==0.100.0
kafka-python==2.0.2
pdf2image==1.16.3
pypdf==3.12.1
pytesseract==0.3.10
python-multipart==0.0.6
uvicorn==0.22.0